Pass around C pointers as Java long
so we're 64bit-ready

+++ library/src/com/litl/leveldb/Iterator.java
@@ -1,12 +1,12 @@
 package com.litl.leveldb;
 
 public class Iterator extends NativeObject {
-    Iterator(int iterPtr) {
+    Iterator(long iterPtr) {
         super(iterPtr);
     }
 
     @Override
-    public void closeNativeObject(int ptr) {
+    public void closeNativeObject(long ptr) {
         nativeDestroy(ptr);
     }
 
@@ -48,19 +48,19 @@ public class Iterator extends NativeObject {
         return nativeValue(mPtr);
     }
 
-    private static native void nativeDestroy(int ptr);
+    private static native void nativeDestroy(long ptr);
 
-    private static native void nativeSeekToFirst(int ptr);
+    private static native void nativeSeekToFirst(long ptr);
 
-    private static native void nativeSeek(int ptr, byte[] key);
+    private static native void nativeSeek(long ptr, byte[] key);
 
-    private static native boolean nativeValid(int ptr);
+    private static native boolean nativeValid(long ptr);
 
-    private static native void nativeNext(int ptr);
+    private static native void nativeNext(long ptr);
 
-    private static native void nativePrev(int ptr);
+    private static native void nativePrev(long ptr);
 
-    private static native byte[] nativeKey(int dbPtr);
+    private static native byte[] nativeKey(long dbPtr);
 
-    private static native byte[] nativeValue(int dbPtr);
+    private static native byte[] nativeValue(long dbPtr);
 }
